pyproject.toml
[tool.poetry]
name = "agora"
version = "0.1.0"
description = "Motor narrativo conversacional para videojuego basado en agentes"
authors = ["Agora Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = "^0.70.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
psycopg = {extras = ["binary"], version = "^3.3.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
agora = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
